Randomization menu option, various fixes
Added menu option for true randomization instead of weighted randomization (the latter of which falls flat due to the limited number of base addons and addon types included with MHUDU). The functionality was present but I completely forgot to make an actual user-accessible menu option for it.
Adjusted DOOM HUD and Shigeru Miyamoto to the bottom of the screen. That'll teach me to test things in windowed mode
Fixed Rocket League addon name/description being labeled as the Clippy addon due to a hasty copypastey
Removed unused menumanager.lua and menu/options.json files
Version bump to 0.0.3